+++ reports/src/main/java/timeboard/reports/ReportKPIController.java
@@ -46,6 +46,7 @@ import timeboard.core.model.Account;
 import timeboard.core.model.Project;
 import timeboard.core.model.Report;
 import timeboard.core.ui.UserInfo;
+
 import java.io.IOException;
 import java.util.List;
 import java.util.concurrent.atomic.AtomicReference;
