+++ reports/src/test/java/timeboard/reports/TestExpressionParser.java
@@ -1,5 +1,31 @@
 package timeboard.reports;
 
+/*-
+ * #%L
+ * reports
+ * %%
+ * Copyright (C) 2019 Timeboard
+ * %%
+ * Permission is hereby granted, free of charge, to any person obtaining a copy
+ * of this software and associated documentation files (the "Software"), to deal
+ * in the Software without restriction, including without limitation the rights
+ * to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+ * copies of the Software, and to permit persons to whom the Software is
+ * furnished to do so, subject to the following conditions:
+ * 
+ * The above copyright notice and this permission notice shall be included in
+ * all copies or substantial portions of the Software.
+ * 
+ * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+ * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+ * FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+ * AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+ * LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+ * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
+ * THE SOFTWARE.
+ * #L%
+ */
+
 import org.junit.Assert;
 import org.junit.Test;
 import org.springframework.expression.Expression;
@@ -12,7 +38,7 @@ public class TestExpressionParser {
     @Test
     public void testParseTrue(){
         final String filter = "tagKey == 'CUSTOMER' && tagValue == 'TIMEBOARD'";
-        final ReportsRestAPI.TagWrapper tag = new ReportsRestAPI.TagWrapper("CUSTOMER","TIMEBOARD");
+        final ReportsService.TagWrapper tag = new ReportsService.TagWrapper("CUSTOMER","TIMEBOARD");
         final ExpressionParser parser = new SpelExpressionParser();
         final Expression exp = parser.parseExpression(filter);
 
@@ -22,7 +48,7 @@ public class TestExpressionParser {
     @Test
     public void testParseFalse(){
 
-        final ReportsRestAPI.TagWrapper tag = new ReportsRestAPI.TagWrapper("CUSTOMER","TIMEBOARD");
+        final ReportsService.TagWrapper tag = new ReportsService.TagWrapper("CUSTOMER","TIMEBOARD");
         final ExpressionParser parser = new SpelExpressionParser();
         final Expression exp = parser.parseExpression("tagKey == 'CUSTOMER' && tagValue != 'TIMEBOARD'");
 
@@ -32,7 +58,7 @@ public class TestExpressionParser {
     @Test(expected = SpelParseException.class)
     public void testParseFail(){
 
-        final ReportsRestAPI.TagWrapper tag = new ReportsRestAPI.TagWrapper("CUSTOMER","TIMEBOARD");
+        final ReportsService.TagWrapper tag = new ReportsService.TagWrapper("CUSTOMER","TIMEBOARD");
         final ExpressionParser parser = new SpelExpressionParser();
         final Expression exp = parser.parseExpression("tagKey == 'CUSTOMER' &@zfez& tagValue != 'TIMEBOARD'");
